Class Establishment
Created the basic variables and classes for Pet, Client, and costs.
Created enums for managing types of costs and how they are quantified and made several basic ones for experimentation.
Unsure how to proceede with invoicing and will deal with it soon, it has been commented out in the client class.
The new package entitled database is just from experimenting with better organization but everything seems to need to be together right now or it cant comunicate so they all got moved into the same one.